new Data() method returns the binary instead of writing to file
This patch turns Write() into a simple convenience wrapper for Data(),
writing to a file. There should be no backward compatibility issues,
but it will let you provide front-end API calls that will return
the data instead of writing it to a file, which could be useful for
web services and whatnot.

WARNING: There is a tradeoff in that it will need to store the entire
PDB in memory before writing it to disk. If this is a problem then I
suggest not taking this patch.